inference/pyproject.toml
[project]
name = "animecraft-inference"
version = "0.1.0"
description = "gRPC inference service for Anime Craft"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
    "Pillow>=11.0.0",
    "numpy>=2.0.0",
    "accelerate>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "grpcio-testing>=1.68.0",
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
protoc = """python -m grpc_tools.protoc \
    -I../proto \
    --python_out=src/animecraft_inference/generated \
    --grpc_python_out=src/animecraft_inference/generated \
    ../proto/inference.proto"""
serve = "python -m animecraft_inference.server"
test = "pytest tests/ -v"
